refactor(chat): split transcript activity modules
