+++ src/extensions/default-layout/turn-actions.tsx
@@ -1,0 +1,126 @@
+import { useState } from "react";
+import type { BuiltinComponentProps } from "../../components/A2UIRenderer";
+import type { ChatMessage } from "../../types/a2ui";
+import { tabIsRunning } from "./message-row-state";
+
+function tabCwdFromState(
+  state: Record<string, unknown>,
+  tabId: string | undefined,
+): string | undefined {
+  if (!tabId || !Array.isArray(state.tabs)) return undefined;
+  const tab = state.tabs.find(
+    (candidate): candidate is { id: string; cwd?: string } =>
+      Boolean(
+        candidate &&
+        typeof candidate === "object" &&
+        "id" in candidate &&
+        candidate.id === tabId,
+      ),
+  );
+  return typeof tab?.cwd === "string" ? tab.cwd : undefined;
+}
+
+function branchEventPayload(
+  target: ChatMessage,
+  tabId: string | undefined,
+  state: Record<string, unknown>,
+) {
+  const cwd = target.cwd ?? tabCwdFromState(state, tabId);
+  return {
+    entryId: target.entryId,
+    tabId,
+    ...(cwd ? { cwd } : {}),
+  };
+}
+
+export function TurnBranchActions({
+  target,
+  state,
+  tabId,
+  onEvent,
+}: {
+  target?: ChatMessage;
+  state: Record<string, unknown>;
+  tabId?: string;
+  onEvent?: BuiltinComponentProps["onEvent"];
+}) {
+  const [confirmingRollback, setConfirmingRollback] = useState(false);
+  if (!target?.entryId || !onEvent || tabIsRunning(state, tabId)) return null;
+  return (
+    <div
+      className="ae-turn-branch-actions"
+      onMouseLeave={() => setConfirmingRollback(false)}
+    >
+      {confirmingRollback ? (
+        <>
+          <button
+            type="button"
+            className="ae-turn-branch-btn ae-turn-branch-confirm"
+            onClick={() => {
+              setConfirmingRollback(false);
+              onEvent(
+                "rollback-to-here",
+                branchEventPayload(target, tabId, state),
+              );
+            }}
+          >
+            Confirm rollback
+          </button>
+          <button
+            type="button"
+            className="ae-turn-branch-btn"
+            onClick={() => setConfirmingRollback(false)}
+          >
+            Cancel
+          </button>
+        </>
+      ) : (
+        <>
+          <button
+            type="button"
+            className="ae-turn-branch-btn"
+            aria-label="Rollback this turn"
+            title="Rewind the conversation to this turn"
+            onClick={() => setConfirmingRollback(true)}
+          >
+            <svg
+              viewBox="0 0 16 16"
+              width="13"
+              height="13"
+              aria-hidden="true"
+              focusable="false"
+            >
+              <path d="M5.2 5.1H10a4 4 0 1 1-3.1 6.55" />
+              <path d="M5.2 5.1 7.55 2.8" />
+              <path d="M5.2 5.1 7.55 7.45" />
+            </svg>
+            <span className="ae-turn-branch-label">Rollback</span>
+          </button>
+          <button
+            type="button"
+            className="ae-turn-branch-btn"
+            aria-label="Fork this turn"
+            title="Fork the conversation into a new tab from this turn"
+            onClick={() =>
+              onEvent("fork-to-tab", branchEventPayload(target, tabId, state))
+            }
+          >
+            <svg
+              viewBox="0 0 16 16"
+              width="13"
+              height="13"
+              aria-hidden="true"
+              focusable="false"
+            >
+              <path d="M5 3.25v4.15c0 2.4 1.55 4.1 4.2 4.1H11" />
+              <path d="M8.75 9.2 11 11.5l-2.25 2.3" />
+              <circle cx="5" cy="3.25" r="1.6" />
+              <circle cx="5" cy="12.75" r="1.6" />
+            </svg>
+            <span className="ae-turn-branch-label">Fork</span>
+          </button>
+        </>
+      )}
+    </div>
+  );
+}
